+++ rbx_dom_lua/src/customProperties.lua
@@ -1,3 +1,30 @@
+local ScriptEditorService = game:GetService("ScriptEditorService")
+
+--- A list of `Enum.Material` values that are used for Terrain.MaterialColors
+local TERRAIN_MATERIAL_COLORS = {
+	Enum.Material.Grass,
+	Enum.Material.Slate,
+	Enum.Material.Concrete,
+	Enum.Material.Brick,
+	Enum.Material.Sand,
+	Enum.Material.WoodPlanks,
+	Enum.Material.Rock,
+	Enum.Material.Glacier,
+	Enum.Material.Snow,
+	Enum.Material.Sandstone,
+	Enum.Material.Mud,
+	Enum.Material.Basalt,
+	Enum.Material.Ground,
+	Enum.Material.CrackedLava,
+	Enum.Material.Asphalt,
+	Enum.Material.Cobblestone,
+	Enum.Material.Ice,
+	Enum.Material.LeafyGrass,
+	Enum.Material.Salt,
+	Enum.Material.Limestone,
+	Enum.Material.Pavement,
+}
+
 -- Defines how to read and write properties that aren't directly scriptable.
 -- The reflection database refers to these as having scriptability = "Custom"
 
@@ -13,9 +40,24 @@ return {
 
 			write = function(instance, _, value)
 				local existing = instance:GetAttributes()
+				local didAllWritesSucceed = true
 
-				for key, attr in pairs(value) do
-					instance:SetAttribute(key, attr)
+				for attributeName, attributeValue in pairs(value) do
+					local isNameValid =
+						-- For our SetAttribute to succeed, the attribute name must be
+						-- less than or equal to 100 characters...
+						#attributeName <= 100
+						-- ...must only contain alphanumeric characters, periods, hyphens,
+						-- underscores, or forward slashes...
+						and attributeName:match("[^%w%.%-_/]") == nil
+						-- ... and must not use the RBX prefix, which is reserved by Roblox.
+						and attributeName:sub(1, 3) ~= "RBX"
+
+					if isNameValid then
+						instance:SetAttribute(attributeName, attributeValue)
+					else
+						didAllWritesSucceed = false
+					end
 				end
 
 				for key in pairs(existing) do
@@ -24,7 +66,7 @@ return {
 					end
 				end
 
-				return true
+				return didAllWritesSucceed
 			end,
 		},
 
@@ -56,11 +98,10 @@ return {
 	},
 	LocalizationTable = {
 		Contents = {
-			read = function(instance, key)
+			read = function(instance, _)
 				return true, instance:GetContents()
 			end,
-
-			write = function(instance, key, value)
+			write = function(instance, _, value)
 				instance:SetContents(value)
 				return true
 			end,
@@ -73,22 +114,79 @@ return {
 			end,
 
 			write = function(meshPart: MeshPart, _, meshId: string)
-				task.spawn(function ()
-					-- Use default settings to do this as immediately as possible.
-					local import = InsertService:CreateMeshPartAsync(meshId, Enum.CollisionFidelity.Box, Enum.RenderFidelity.Precise)
-					
-					-- Cache the desired fidelity settings.
-					local collisionFidelity = meshPart.CollisionFidelity
-					local renderFidelity = meshPart.RenderFidelity
+				-- Use default settings to do this as immediately as possible.
+				local import = InsertService:CreateMeshPartAsync(meshId, Enum.CollisionFidelity.Box, Enum.RenderFidelity.Precise)
+				
+				-- Cache the desired fidelity settings.
+				local collisionFidelity = meshPart.CollisionFidelity
+				local renderFidelity = meshPart.RenderFidelity
 
-					-- Apply the mesh.
-					meshPart:ApplyMesh(import)
+				-- Apply the mesh.
+				meshPart:ApplyMesh(import)
 
-					-- Let studio handle the fidelities in serial.
-					meshPart.CollisionFidelity = collisionFidelity
-					meshPart.RenderFidelity = renderFidelity
+				-- Let studio handle the fidelities in serial.
+				meshPart.CollisionFidelity = collisionFidelity
+				meshPart.RenderFidelity = renderFidelity
+			end
+		}
+	},
+	Model = {
+		Scale = {
+			read = function(instance, _, _)
+				return true, instance:GetScale()
+			end,
+			write = function(instance, _, value)
+				return true, instance:ScaleTo(value)
+			end,
+		},
+	},
+	Terrain = {
+		MaterialColors = {
+			read = function(instance: Terrain)
+				-- There's no way to get a list of every color, so we have to
+				-- make one.
+				local colors = {}
+				for _, material in TERRAIN_MATERIAL_COLORS do
+					colors[material] = instance:GetMaterialColor(material)
+				end
+
+				return true, colors
+			end,
+			write = function(instance: Terrain, _, value: { [Enum.Material]: Color3 })
+				for material, color in value do
+					instance:SetMaterialColor(material, color)
+				end
+				return true
+			end,
+		},
+	},
+	Script = {
+		Source = {
+			read = function(instance: Script)
+				return true, ScriptEditorService:GetEditorSource(instance)
+			end,
+			write = function(instance: Script, _, value: string)
+				task.spawn(function()
+					ScriptEditorService:UpdateSourceAsync(instance, function()
+						return value
+					end)
 				end)
-
+				return true
+			end,
+		},
+	},
+	ModuleScript = {
+		Source = {
+			read = function(instance: ModuleScript)
+				return true, ScriptEditorService:GetEditorSource(instance)
+			end,
+			write = function(instance: ModuleScript, _, value: string)
+				task.spawn(function()
+					ScriptEditorService:UpdateSourceAsync(instance, function()
+						return value
+					end)
+				end)
+				
 				return true
 			end,
 		},