+++ rbx_xml/src/deserializer.rs
@@ -8,7 +8,7 @@ use rbx_dom_weak::{
     types::{Ref, SharedString, Variant, VariantType},
     InstanceBuilder, WeakDom,
 };
-use rbx_reflection::DataType;
+use rbx_reflection::{DataType, PropertyKind, PropertySerialization, ReflectionDatabase};
 
 use crate::{
     conversion::ConvertVariant,
@@ -69,16 +69,18 @@ pub enum DecodePropertyBehavior {
 
 /// Options available for deserializing an XML-format model or place.
 #[derive(Debug, Clone)]
-pub struct DecodeOptions {
+pub struct DecodeOptions<'db> {
     property_behavior: DecodePropertyBehavior,
+    database: &'db ReflectionDatabase<'db>,
 }
 
-impl DecodeOptions {
+impl<'db> DecodeOptions<'db> {
     /// Constructs a `DecodeOptions` with all values set to their defaults.
     #[inline]
     pub fn new() -> Self {
         DecodeOptions {
             property_behavior: DecodePropertyBehavior::IgnoreUnknown,
+            database: rbx_reflection_database::get(),
         }
     }
 
@@ -86,7 +88,17 @@ impl DecodeOptions {
     /// ones.
     #[inline]
     pub fn property_behavior(self, property_behavior: DecodePropertyBehavior) -> Self {
-        DecodeOptions { property_behavior }
+        DecodeOptions {
+            property_behavior,
+            ..self
+        }
+    }
+
+    /// Determines what reflection database rbx_xml will use to deserialize
+    /// properties.
+    #[inline]
+    pub fn reflection_database(self, database: &'db ReflectionDatabase<'db>) -> Self {
+        DecodeOptions { database, ..self }
     }
 
     /// A utility function to determine whether or not we should reference the
@@ -96,16 +108,17 @@ impl DecodeOptions {
     }
 }
 
-impl Default for DecodeOptions {
-    fn default() -> DecodeOptions {
+impl<'db> Default for DecodeOptions<'db> {
+    fn default() -> DecodeOptions<'db> {
         DecodeOptions::new()
     }
 }
 
 /// The state needed to deserialize an XML model into an `WeakDom`.
-pub struct ParseState<'a> {
-    tree: &'a mut WeakDom,
-    options: DecodeOptions,
+pub struct ParseState<'dom, 'db> {
+    tree: &'dom mut WeakDom,
+
+    options: DecodeOptions<'db>,
 
     /// Metadata deserialized from 'Meta' fields in the file.
     /// Known fields are:
@@ -150,8 +163,8 @@ struct SharedStringRewrite {
     shared_string_hash: String,
 }
 
-impl<'a> ParseState<'a> {
-    fn new(tree: &mut WeakDom, options: DecodeOptions) -> ParseState {
+impl<'dom, 'db> ParseState<'dom, 'db> {
+    fn new(tree: &'dom mut WeakDom, options: DecodeOptions<'db>) -> ParseState<'dom, 'db> {
         ParseState {
             tree,
             options,
@@ -559,7 +572,11 @@ fn deserialize_properties<R: Read>(
         );
 
         let maybe_descriptor = if state.options.use_reflection() {
-            find_canonical_property_descriptor(&class_name, &xml_property_name)
+            find_canonical_property_descriptor(
+                &class_name,
+                &xml_property_name,
+                state.options.database,
+            )
         } else {
             None
         };
@@ -588,6 +605,7 @@ fn deserialize_properties<R: Read>(
                 DataType::Enum(_enum_name) => VariantType::Enum,
                 _ => unimplemented!(),
             };
+            log::trace!("property's read type: {xml_ty:?}, canonical type: {expected_type:?}");
 
             let value = match value.try_convert(expected_type) {
                 Ok(value) => value,
@@ -607,7 +625,34 @@ fn deserialize_properties<R: Read>(
                 }
             };
 
-            props.insert(descriptor.name.to_string(), value);
+            match &descriptor.kind {
+                PropertyKind::Canonical {
+                    serialization: PropertySerialization::Migrate(migration),
+                } => {
+                    let new_property_name = &migration.new_property_name;
+                    let old_property_name = &descriptor.name;
+
+                    if !props.contains_key(new_property_name) {
+                        log::trace!(
+                            "Attempting to migrate property {old_property_name} to {new_property_name}"
+                        );
+                        match migration.perform(&value) {
+                            Ok(migrated_value) => {
+                                props.insert(new_property_name.to_string(), migrated_value);
+                                log::trace!(
+                                    "Successfully migrated property {old_property_name} to {new_property_name}"
+                                );
+                            }
+                            Err(error) => {
+                                return Err(reader.error(DecodeErrorKind::MigrationError(error)));
+                            }
+                        }
+                    }
+                }
+                _ => {
+                    props.insert(descriptor.name.to_string(), value);
+                }
+            };
         } else {
             match state.options.property_behavior {
                 DecodePropertyBehavior::IgnoreUnknown => {