+++ rbx_xml/src/serializer.rs
@@ -1,4 +1,5 @@
 use std::{
+    borrow::Cow,
     collections::{BTreeMap, HashMap},
     io::Write,
 };
@@ -7,7 +8,7 @@ use rbx_dom_weak::{
     types::{Ref, SharedString, SharedStringHash, Variant, VariantType},
     WeakDom,
 };
-use rbx_reflection::DataType;
+use rbx_reflection::{DataType, PropertyKind, PropertySerialization, ReflectionDatabase};
 
 use crate::{
     conversion::ConvertVariant,
@@ -73,16 +74,18 @@ pub enum EncodePropertyBehavior {
 
 /// Options available for serializing an XML-format model or place.
 #[derive(Debug, Clone)]
-pub struct EncodeOptions {
+pub struct EncodeOptions<'db> {
     property_behavior: EncodePropertyBehavior,
+    database: &'db ReflectionDatabase<'db>,
 }
 
-impl EncodeOptions {
+impl<'db> EncodeOptions<'db> {
     /// Constructs a `EncodeOptions` with all values set to their defaults.
     #[inline]
     pub fn new() -> Self {
         EncodeOptions {
             property_behavior: EncodePropertyBehavior::IgnoreUnknown,
+            database: rbx_reflection_database::get(),
         }
     }
 
@@ -90,7 +93,17 @@ impl EncodeOptions {
     /// ones.
     #[inline]
     pub fn property_behavior(self, property_behavior: EncodePropertyBehavior) -> Self {
-        EncodeOptions { property_behavior }
+        EncodeOptions {
+            property_behavior,
+            ..self
+        }
+    }
+
+    /// Determines what reflection database rbx_xml will use to serialize
+    /// properties.
+    #[inline]
+    pub fn reflection_database(self, database: &'db ReflectionDatabase<'db>) -> Self {
+        EncodeOptions { database, ..self }
     }
 
     pub(crate) fn use_reflection(&self) -> bool {
@@ -98,14 +111,14 @@ impl EncodeOptions {
     }
 }
 
-impl Default for EncodeOptions {
-    fn default() -> EncodeOptions {
+impl<'db> Default for EncodeOptions<'db> {
+    fn default() -> EncodeOptions<'db> {
         EncodeOptions::new()
     }
 }
 
-pub struct EmitState {
-    options: EncodeOptions,
+pub struct EmitState<'db> {
+    options: EncodeOptions<'db>,
 
     /// A map of IDs written so far to the generated referent that they use.
     /// This map is used to correctly emit Ref properties.
@@ -119,8 +132,8 @@ pub struct EmitState {
     shared_strings_to_emit: BTreeMap<SharedStringHash, SharedString>,
 }
 
-impl EmitState {
-    pub fn new(options: EncodeOptions) -> EmitState {
+impl<'db> EmitState<'db> {
+    pub fn new(options: EncodeOptions<'db>) -> EmitState<'db> {
         EmitState {
             options,
             referent_map: HashMap::new(),
@@ -150,12 +163,12 @@ impl EmitState {
 ///
 /// `property_buffer` is a Vec that can be reused between calls to
 /// serialize_instance to make sorting properties more efficient.
-fn serialize_instance<'a, W: Write>(
+fn serialize_instance<'dom, W: Write>(
     writer: &mut XmlEventWriter<W>,
     state: &mut EmitState,
-    tree: &'a WeakDom,
+    tree: &'dom WeakDom,
     id: Ref,
-    property_buffer: &mut Vec<(&'a String, &'a Variant)>,
+    property_buffer: &mut Vec<(&'dom String, &'dom Variant)>,
 ) -> Result<(), NewEncodeError> {
     let instance = tree.get_by_ref(id).unwrap();
     let mapped_id = state.map_id(id);
@@ -182,7 +195,11 @@ fn serialize_instance<'a, W: Write>(
 
     for (property_name, value) in property_buffer.drain(..) {
         let maybe_serialized_descriptor = if state.options.use_reflection() {
-            find_serialized_property_descriptor(&instance.class, property_name)
+            find_serialized_property_descriptor(
+                &instance.class,
+                property_name,
+                state.options.database,
+            )
         } else {
             None
         };
@@ -194,7 +211,9 @@ fn serialize_instance<'a, W: Write>(
                 _ => unimplemented!(),
             };
 
-            let converted_value = match value.try_convert_ref(data_type) {
+            let mut serialized_name = serialized_descriptor.name.as_ref();
+
+            let mut converted_value = match value.try_convert_ref(data_type) {
                 Ok(value) => value,
                 Err(message) => {
                     return Err(
@@ -209,7 +228,20 @@ fn serialize_instance<'a, W: Write>(
                 }
             };
 
-            write_value_xml(writer, state, &serialized_descriptor.name, &converted_value)?;
+            // Perform migrations during serialization
+            if let PropertyKind::Canonical {
+                serialization: PropertySerialization::Migrate(migration),
+            } = &serialized_descriptor.kind
+            {
+                // If the migration fails, there's no harm in us doing nothing
+                // since old values will still load in Studio.
+                if let Ok(new_value) = migration.perform(&converted_value) {
+                    converted_value = Cow::Owned(new_value);
+                    serialized_name = &migration.new_property_name
+                }
+            }
+
+            write_value_xml(writer, state, serialized_name, &converted_value)?;
         } else {
             match state.options.property_behavior {
                 EncodePropertyBehavior::IgnoreUnknown => {}